Restore normalized URLs
Troxy supports a nested URL format, e.g.
'http://troxyhostname:8083/https://the.actual.host/somepath', where the
path contains the actual URL the application intends to use. However,
this often leads to problems with various libraries and how they handle
URls containing double-slash ('//'). Many of them will try to normalize
the URLs by removing a '/', which breaks the 'protocol-part' of the URL.

We remedy this problem by trying to restore these normalized URLs. If the
path contains 'http(s):/', we assume that a '/' was removed and we insert
a new one.