Matches headers in AutoRegexNewRecording line by line instead of entire input at once
AutoRegexNewRecording currently matches an entire request header against one, gigantic RegEx which is created by concatenating all RegExes in the config. This makes it hard to reason about how one RegEx may behave when concatenated with several others.

This commit changes how AutoRegexNewRecording to matching each pattern separately, line by line. Since the format of a HTTP header is well defined, this should not cause any issues. As a nice side effect, the RegExes specified in the config no longer have to deal with newlines. The changes to AutoRegexNewRecording do not change how the cache is searched, so existing recordings should continue to work without problems.